+++ iodevices/proctec-vdm.c
@@ -15,6 +15,7 @@
  * 04-NOV-2019 eliminate usage of mem_base()
  */
 
+#include <stdint.h>
 #include <X11/X.h>
 #include <X11/Xlib.h>
 #include <X11/Xutil.h>
@@ -209,9 +210,9 @@ static void refresh(void)
 /* thread for updating the display */
 static void *update_display(void *arg)
 {
-	extern unsigned long long get_clock_us(void);
+	extern uint64_t get_clock_us(void);
 
-	unsigned long long t1, t2;
+	uint64_t t1, t2;
 	int tdiff;
 
 	UNUSED(arg);
